Hack out certain things so it runs
This commit comments out refrences to controls that could not be found
when I tried to run the application after a default installation of
Visual Basic 5 on Windows 10.

The application will launch and seems to work a bit!

What I would rather do is figure out which controls the application is
missing, which I believe can be found in the logs, and add them back in.